Smoothed out the builder functionality.  Added the start of other types of children (models, behaviors, and validators).  Models are currently implemented, with a handful of types in there.  These can also be dynamically overridden (like Components).
The configuration and general interface needs to be cleaned up, but it is functional and *should* be usable in an actual product soon.

Also added a convenience feature for development.  The builder constructs a string of suggested HTML code for the givent Component hierarchy, which can be cut and pasted into the HTML file for the builder.

git-svn-id: https://wicket-stuff.svn.sourceforge.net/svnroot/wicket-stuff/branches/wicket-1.3/wicket-contrib-groovy@1849 ef7698a4-5110-0410-9fc6-c7eb3693863f